requirements.txt
crewai
crewai[tools]
openai
pyairtable
requests
praw
python-dotenv
tldextract
beautifulsoup4
requests
